requirements.txt
Django==2.2.1
django-tinymce4-lite==1.7.5
jsmin==2.2.2
Pillow==6.0.0
psycopg2==2.8.2
pytz==2019.1
sqlparse==0.3.0
